refactor: cleanup Doctrine Tracing Middleware
