--- src/Middleware/Doctrine/Metric/Connection.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\Middleware\Doctrine\Metric;
-
-use Doctrine\DBAL\Driver\Middleware\AbstractConnectionMiddleware;
-
-class Connection extends AbstractConnectionMiddleware
-{
-}
