--- src/Middleware/Doctrine/Metric/Driver.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\Middleware\Doctrine\Metric;
-
-use Doctrine\DBAL\Driver\Middleware\AbstractDriverMiddleware;
-
-final class Driver extends AbstractDriverMiddleware
-{
-}
