--- src/Middleware/Doctrine/Metric/Middleware.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\Middleware\Doctrine\Metric;
-
-use Doctrine\DBAL\Driver as DriverInterface;
-use Doctrine\DBAL\Driver\Middleware as MiddlewareInterface;
-
-class Middleware implements MiddlewareInterface
-{
-    public function wrap(DriverInterface $driver): DriverInterface
-    {
-        return new Driver($driver);
-    }
-}
