--- src/Middleware/Doctrine/Metric/Statement.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\Middleware\Doctrine\Metric;
-
-use Doctrine\DBAL\Driver\Middleware\AbstractStatementMiddleware;
-
-final class Statement extends AbstractStatementMiddleware
-{
-}
