+++ src/Middleware/Doctrine/Trace/Statement.php
@@ -5,14 +5,11 @@ namespace FriendsOfOpenTelemetry\OpenTelemetryBundle\Middleware\Doctrine\Trace;
 use Doctrine\DBAL\Driver\Middleware\AbstractStatementMiddleware;
 use Doctrine\DBAL\Driver\Result;
 use Doctrine\DBAL\Driver\Statement as StatementInterface;
-use OpenTelemetry\API\Trace\TracerInterface;
 
 final class Statement extends AbstractStatementMiddleware
 {
     public function __construct(
-        private StatementInterface $statement,
-        private TracerInterface $tracer,
-        private string $sql,
+        StatementInterface $statement,
     ) {
         parent::__construct($statement);
     }
